Title: 숫자들의 합이 7의 배수, Time: 835ms, Memory: 1MB, Status: Passed - Codetree

+++ 250719/숫자들의 합이 7의 배수/the-sum-of-the-numbers-is-a-multiple-of-7.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int n;
 int numbers[50001];
-int prefix[50001];
+long long prefix[50001];
 
 
 int main() {
